+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## dense 1.1.7 not yet released
 
-* Implement Dense.flatten(hash)
+* Implement Dense.deflate(hash)
 
 
 ## dense 1.1.6 released 2018-12-24  (Merry Christmas!)